requirements.txt
faiss_cpu==1.8.0
fastapi==0.110.0
numpy==1.26.4
Pillow==10.2.0
pydantic==2.6.4
torch==2.2.1
transformers==4.38.2
